Move process-wrangling classes to Functions lib
This is a straight R#-enabled move of the classes to the new project,
with namespaces updated along the way.
