+++ Solutions/Corvus.AzureFunctions/Corvus/AzureFunctions/Internal/FunctionConfiguration.cs
