+++ Solutions/Corvus.AzureFunctions/Corvus/AzureFunctions/Internal/FunctionOutputBufferHandler.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions.Internal
+namespace Corvus.AzureFunctions.Internal
 {
     using System.Diagnostics;
     using System.Threading.Tasks;
