+++ Solutions/Corvus.AzureFunctions/Corvus/AzureFunctions/Internal/ProcessOutputHandler.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions.Internal
+namespace Corvus.AzureFunctions.Internal
 {
     using System;
     using System.Diagnostics;
