+++ Solutions/Corvus.SpecFlow.Extensions.Demo/AzureFunctionsTesting/StepBindings.cs
@@ -8,7 +8,7 @@ namespace Corvus.SpecFlow.Extensions.Demo.AzureFunctionsTesting
     using System.Net.Http;
     using System.Text;
     using System.Threading.Tasks;
-    using Corvus.SpecFlow.Extensions;
+    using Corvus.AzureFunctions;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
     using NUnit.Framework;
